requirements.txt
kiteconnect
redis
pandas_ta